Split properties out of core.py
