+++ src/invent/ui/export.py
@@ -94,6 +94,7 @@ experimental_create_proxy = "auto"
 "{{INVENT}}/ui/core.py"="{{INVENT_TO}}/ui/core.py"
 "{{INVENT}}/ui/export.py"="{{INVENT_TO}}/ui/export.py"
 "{{INVENT}}/ui/page.py"="{{INVENT_TO}}/ui/page.py"
+"{{INVENT}}/ui/properties.py"="{{INVENT_TO}}/ui/properties.py"
 "{{INVENT}}/ui/utils.py"="{{INVENT_TO}}/ui/utils.py"
 "{{INVENT}}/ui/widgets/__init__.py"="{{INVENT_TO}}/ui/widgets/__init__.py"
 "{{INVENT}}/ui/widgets/audio.py"="{{INVENT_TO}}/ui/widgets/audio.py"
