+++ src/invent/ui/widgets/audio.py
@@ -18,11 +18,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from invent.ui.core import (
-    Widget,
-    TextProperty,
-    MessageBlueprint,
-)
+from invent.ui.core import Widget, MessageBlueprint
+from invent.ui.properties import TextProperty
 from pyscript import document
