+++ src/invent/ui/widgets/button.py
@@ -18,14 +18,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from invent.ui.core import (
-    Widget,
-    TextProperty,
-    BooleanProperty,
-    ChoiceProperty,
-    MessageBlueprint,
-)
-
+from invent.ui.core import Widget, MessageBlueprint
+from invent.ui.properties import TextProperty, BooleanProperty, ChoiceProperty
 from pyscript import document
 from pyscript.ffi import create_proxy
