+++ src/invent/ui/widgets/textinput.py
@@ -21,13 +21,12 @@ limitations under the License.
 from pyscript import document
 from pyscript.ffi import create_proxy
 
-from invent.ui.core import (
-    Widget,
+from invent.ui.core import Widget, MessageBlueprint
+from invent.ui.properties import (
     TextProperty,
     IntegerProperty,
     BooleanProperty,
     ChoiceProperty,
-    MessageBlueprint,
 )
